Separate surface functions into a module (#412)
Also updated logging to write all stderr stdout to log too by default
