+++ hippunfold/workflow/lib/utils.py
@@ -10,12 +10,18 @@ def setup_logger(log_file=None):
 
     formatter = logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
 
-    handler = (
-        logging.FileHandler(log_file) if log_file else logging.StreamHandler(sys.stderr)
-    )
+    if log_file:
+        handler = logging.FileHandler(log_file)
+    else:
+        handler = logging.StreamHandler(sys.stderr)
     handler.setFormatter(formatter)
-
     logger.addHandler(handler)
+
+    if log_file:
+        # Redirect stdout and stderr to the same log file
+        sys.stdout = open(log_file, "a")
+        sys.stderr = open(log_file, "a")
+
     return logger
