+++ hippunfold/workflow/scripts/gen_isosurface.py
@@ -1,9 +1,13 @@
 import pyvista as pv
 import nibabel as nib
 import numpy as np
-from copy import deepcopy
 from scipy.ndimage import binary_dilation
 from lib.utils import setup_logger
+from lib.surface import (
+    write_surface_to_gifti,
+    apply_affine_transform,
+    remove_nan_vertices,
+)
 
 log_file = snakemake.log[0] if snakemake.log else None
 logger = setup_logger(log_file)
@@ -29,58 +33,6 @@ def get_adjacent_voxels(mask_a, mask_b):
     )
 
     return adjacency_mask
-
-
-def write_surface_to_gifti(mesh, out_surf_gii):
-    """
-    Writes a PyVista mesh to a GIFTI surface file.
-    """
-    points = mesh.points
-    faces = mesh.faces.reshape((-1, 4))[:, 1:4]  # Extract triangle indices
-
-    points_darray = nib.gifti.GiftiDataArray(
-        data=points, intent="NIFTI_INTENT_POINTSET", datatype="NIFTI_TYPE_FLOAT32"
-    )
-
-    tri_darray = nib.gifti.GiftiDataArray(
-        data=faces, intent="NIFTI_INTENT_TRIANGLE", datatype="NIFTI_TYPE_INT32"
-    )
-
-    gifti = nib.GiftiImage()
-    gifti.add_gifti_data_array(points_darray)
-    gifti.add_gifti_data_array(tri_darray)
-    gifti.to_filename(out_surf_gii)
-
-
-def apply_affine_transform(mesh, affine_matrix, inverse=False):
-    """
-    Applies an affine transformation to a PyVista mesh.
-    """
-    if inverse:
-        affine_matrix = np.linalg.inv(affine_matrix)
-
-    transformed_points = (
-        np.hstack([mesh.points, np.ones((mesh.n_points, 1))]) @ affine_matrix.T
-    )[:, :3]
-    transformed_mesh = pv.PolyData(transformed_points, mesh.faces)
-    return transformed_mesh
-
-
-def remove_nan_vertices(mesh):
-    """
-    Removes NaN vertices from a PyVista mesh and updates faces accordingly.
-    """
-    valid_mask = ~np.isnan(mesh.points).any(axis=1)
-    new_indices = np.full(mesh.n_points, -1, dtype=int)
-    new_indices[valid_mask] = np.arange(valid_mask.sum())
-
-    faces = mesh.faces.reshape((-1, 4))[:, 1:4]  # Extract triangle indices
-    valid_faces_mask = np.all(valid_mask[faces], axis=1)
-    new_faces = new_indices[faces[valid_faces_mask]]
-    new_faces_pv = np.hstack([np.full((new_faces.shape[0], 1), 3), new_faces])
-
-    cleaned_mesh = pv.PolyData(mesh.points[valid_mask], new_faces_pv)
-    return cleaned_mesh
 
 
 # Load data
@@ -129,18 +81,6 @@ logger.info("Extracting largest connected component")
 surface = surface.extract_largest()
 logger.info(surface)
 
-"""
-#still experimenting with this..
-
-logger.info("Applying surface smoothing")  
-surface = surface.smooth_taubin(#normalize_coordinates=True,
-                                #boundary_smoothing=True,
-                                #feature_smoothing=True,
-                                n_iter=20,
-                                pass_band=0.1)  #n_iter=30, pass_band=0.1)
-logger.info(surface)
-"""
-
 logger.info(f"Filling holes up to radius {snakemake.params.hole_fill_radius}")
 surface = surface.fill_holes(snakemake.params.hole_fill_radius)
 logger.info(surface)
@@ -150,10 +90,6 @@ logger.info(f"Decimating surface with {snakemake.params.decimate_opts}")
 surface = surface.decimate(**snakemake.params.decimate_opts)
 logger.info(surface)
 
-
-# logger.info("applying surface smoothing") # -- still experimenting with smoothing before/after decimation
-# surface = surface.smooth_taubin()
-# logger.info(surface)
 
 logger.info(f"final surface clean to remove overlapping vertices, etc.")
 surface = surface.clean()
